djangobackend/requirements.txt
asgiref==3.2.10
branca==0.3.1
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
Django==3.1.1
django-filter==2.3.0
django-rest-swagger==2.2.0
djangorestframework==3.11.1
folium==0.10.1
idna==2.8
importlib-metadata==1.7.0
itypes==1.2.0
Jinja2==2.10.3
kiwisolver==1.1.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.1.2
mysql-connector-python==8.0.21
mysqlclient==2.0.1
numpy==1.17.3
openapi-codec==1.3.2
pandas==0.25.3
protobuf==3.13.0
pyparsing==2.4.5
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
scipy==1.3.1
seaborn==0.9.0
simplejson==3.17.2
six==1.13.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.7
zipp==3.1.0
